# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Peak Academy API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
...
  /api/sessions:
    get:
      summary: List sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessions list
    post:
      summary: Create teacher session
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Session created
...
  /api/questions:
    post:
      summary: Ask question
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Question created
    get:
      summary: List unanswered questions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Questions list
...
  /api/quizzes:
    post:
      summary: Create live quiz
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Quiz created
  /api/quizzes/{quizId}:
    get:
      summary: Get quiz with questions
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: quizId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quiz details
  /api/quizzes/{quizId}/questions:
    post:
      summary: Add quiz question
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: quizId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Question added
  /api/quizzes/{quizId}/start:
    post:
      summary: Start quiz
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: quizId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quiz started
  /api/quizzes/{quizId}/submit:
    post:
      summary: Submit quiz answers
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: quizId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Attempt submitted
  /api/quizzes/{quizId}/analytics:
    get:
      summary: Quiz analytics
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: quizId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Analytics data
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT